Move type style mapping to utils.
